refactor(theme): rename SiteChrome to SiteLayout
